HotelBillingSystem - Copy/requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.2.0
autobahn==22.2.2
Automat==20.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
constantly==15.1.0
crispy-bootstrap5==0.6
cryptography==36.0.0
cssselect2==0.4.1
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.7
django-ajax-selects==2.1.0
django-allauth==0.38.0
django-appconf==1.0.4
django-bootstrap-v5==1.0.10
django-compressor==2.2
django-crispy-forms==1.13.0
django-filter==21.1
django-gst-field==0.0.3
django-heroku==0.3.1
django-image-cropping==1.6.1
django-imagekit==4.1.0
django-libsass==0.7
django-model-utils==4.1.1
django-phonenumber-field==5.2.0
django-sass-processor==1.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
easy-thumbnails==2.8
Faker==9.8.0
fontawesome-free==5.15.4
fontawesomefree==6.1.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
libsass==0.21.0
lxml==4.7.1
Markdown==3.3.6
msgpack==1.0.3
msgpack-python==0.5.6
numpy==1.22.3
oauthlib==3.2.0
outcome==1.1.0
packaging==21.3
pdfkit==1.0.0
phonenumbers==8.12.36
pilkit==2.0
Pillow==8.3.2
playwright==1.17.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyee==8.2.2
pyOpenSSL==21.0.0
pyparsing==3.0.7
PyPDF2==1.26.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
rcssmin==1.0.6
redis==4.1.4
reportlab==3.6.5
requests==2.26.0
requests-oauthlib==1.3.1
rjsmin==1.0.12
selenium==4.1.3
selenium-stealth==1.0.6
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
svglib==1.1.0
swapper==1.1.2.post1
text-unidecode==1.3
TikTokApi==4.0.4
tinycss2==1.1.1
tqdm==4.62.2
trio==0.19.0
trio-websocket==0.9.2
Twisted==22.2.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.7
uuid==1.30
webdriver-manager==3.5.4
webencodings==0.5.1
websockets==10.1
whitenoise==5.3.0
wrapt==1.14.0
wsproto==1.0.0
zope.interface==5.4.0
